Fix: reset_demo() apagava tudo e nunca recarregava o seed
- reset_demo() só truncava as tabelas quando o app estava online (Supabase),
  nunca reinseria os dados — clicar em "Reiniciar demonstração" em produção
  deixava categorias, produtos, clientes, ofertas e resgates vazios
- A função agora embute o seed completo (24 clientes, ~50 produtos, ofertas,
  resgates, pedidos históricos), então reiniciar sempre volta pro estado
  completo, não uma base vazia
- Corrigido também um bug latente no seed original: round(double precision, 2)
  não existe no Postgres, faltava cast pra numeric

+++ supabase/migrations/0002_seed.sql
@@ -197,8 +197,8 @@ begin
       criado := now() - ((7 + (random() * 18)::int) || ' days')::interval;
       update public.clientes set criado_em = criado where id = cid;
       itens := jsonb_build_array(
-        jsonb_build_object('produtoId','p-picanha','pesoKg',round((0.5 + random()*2.5)::numeric,2),'preparos','[]'::jsonb,'precoUnitarioAplicado',89.9,'subtotal',round((0.5 + random()*2.5) * 89.9, 2)),
-        jsonb_build_object('produtoId','p-linguica-toscana','pesoKg',round((0.3 + random()*1)::numeric,2),'preparos','["Peça inteira"]'::jsonb,'precoUnitarioAplicado',29.9,'subtotal',round((0.3 + random()*1) * 29.9, 2))
+        jsonb_build_object('produtoId','p-picanha','pesoKg',round((0.5 + random()*2.5)::numeric,2),'preparos','[]'::jsonb,'precoUnitarioAplicado',89.9,'subtotal',round(((0.5 + random()*2.5) * 89.9)::numeric, 2)),
+        jsonb_build_object('produtoId','p-linguica-toscana','pesoKg',round((0.3 + random()*1)::numeric,2),'preparos','["Peça inteira"]'::jsonb,'precoUnitarioAplicado',29.9,'subtotal',round(((0.3 + random()*1) * 29.9)::numeric, 2))
       );
       subtot := (itens->0->>'subtotal')::numeric + (itens->1->>'subtotal')::numeric;
       cb := round(subtot * 0.035, 2);
